Add a documentation section on storing backups in immutable storage
This commit adds a section on how a barman server has to be
configured to store the backups and WALs to a WORM storage. It
covers configuration options with instrcutions on how they should
be set.

There's also a fix of a typo in the enumeration and a mistake in a
note related to incremental rsync backups.

References: BAR-534

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>