Fix behavior when using commands with shortcuts `last-full`/`latest-full`
The shortcuts last-full / latest-full were retrieving not the last full backup
of the server, but the last full backup of the server which was eligible as the
parent for an incremental backup.

While this was the expected behavior, the feedback from the community has shown
that it was confusing for the users.

As a result of the feedback received, the shortcuts will retrieve the last full
backup completed (backup_methods `rsync` or `postgres`) of the Barman server,
without taking into account if it is eligible as a parent for incremental backups.

The eligibility as the parent for incremental backups will still be checked later
in the process, and a proper message will be displayed in case it doesn't suit as
a parent.

Note: the method `is_full_and_eligible_for_incremental`, which is used by the
shortcuts under the hood, has been renamed to a `is_full` property, to better
reflect its behavior. Besides that, we added those two shortcuts to the list of
reserved backup IDs, as that was an oversight in the original implementation.

References: BAR-555

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>