Add support for `DefaultAzureCredential` option on Azure authentication
This adds the `default` option to `--azure-credential` in
`barman-cloud-*` and to `azure_credential` in the server
configuration, allowing the user to use `DefaultAzureCredential`
directly. Previously, it would only be used as a fallback when
no credential was provided and none of the environment variables
were set (AZURE_STORAGE_CONNECTION_STRING, AZURE_STORAGE_SAS_TOKEN,
and AZURE_STORAGE_KEY).

References: BAR-539

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>