Change docs related to restore operation without a backup_id
This commit changes the docs related to the barman restore and
barman-cloud-restore commands.

Both commands now have the `backup_id` option as required (as it was
before Barman 3.13) and a new "auto" value was added. This new value
was introduced so Barman can automatically find suitable backups
without the need to pass a specific `backup_id`.

A note was added in the backup command section to inform the user
about reserverd words when naming a new backup.

References: BAR-596

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>