Add --if-not-exists flag to receive-wal
Adds a new flag `--if-not-exists` to `barman receive-wal` command.

This flag prevents `receive-wal` from exiting with an error when
a slot with the same name already exists. This is the same behavior
as calling `pg_receivewal` with the same flag. This flag will have
an effect only with `--create-slot`, otherwise it is just ignored.

References: BAR-527

Special thanks to @crazybolillo <antonio@zoftko.com> a.k.a @antonag32
for creating this PR and contributing to Barman.

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>