Add compression level in Barman core
Adds a new `compression_level` parameter to the Barman configuration.
This option accepts any integer value or one of the predefined options
`fast`, `average` and `high`, each option corresponding to a different
level depending on the compression algorithm chosen.

Validators/parsers were also added to the new `compression_level`
parameter and to the existing one `compression`, as it was missing
validation.

References: BAR-540

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>