Refactor cloud WAL compression to reuse barman core logic
Some logic of the `barman.clients.cloud_compression` module was moved to
the `barman.compression` module so that we start centralizing the
compression logic in a single place and be able to reuse it across
places. This avoids reimplementing stuff twice whenever we need to modify
compression-related things in Barman.

As a consequence, the Snappy algorithm, previously only available in
`barman-cloud-wal-archive`, is now also available in Barman server.

References: BAR-557

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>