Fix argument parsing issue in `barman restore` and `barman-cloud-restore`
In Barman 3.13.0, a bug was introduced in `barman restore` and
`barman-cloud-restore` due to the handling of intermixed positional and optional
keyword arguments. Specifically, the change made the `backup_id` argument
optional, but the `argparse` module does not allow optional keyword arguments to
be mixed with optional positional arguments.

This caused the following error, which has been reported by the CloudNativePG
project through #6943 in their GitHub repository:

    barman-cloud-restore: error: unrecognized arguments: recovery_dir

When attempting to run a command like this:

    barman-cloud-restore source_url server_name backup_id --cloud-provider aws-s3 recovery_dir

In the case of `barman-cloud-restore`, the issue occurs because `argparse` treats
`--cloud-provider` as the last positional argument, leaving `recovery_dir` unrecognized.
A similar issue can be faced in `barman restore`, when mixing optional keyword arguments
with optional positional arguments.

To fix this:

* The `backup_id` is now required again, and an "auto" option was added to let users
  restore from an automatically chosen backup without specifying a `backup_id`.
* This change was applied to both the `barman-cloud-restore` and `barman restore`
  commands to handle the argument parsing correctly.

References: BAR-596

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>