Create a stub for the `MainRecoveryExecutor` class
This new class is intended to cover the recovery of backups like
these:

* Compressed backup.
* Incremental backup.
* Incremental backup with one or more compressed backup in the
  chain.

The final idea is:

* Use `RecoveryExecutor` to recover a full backup in plain format.
* Use `SnapshotRecoveryExecutor` to recover a snapshot backup.
* Use `MainRecoveryExecutor` for anything else, which relates to
  compressed and/or incremental backups.

With that, we intend to discontinue the `IncrementalRecoveryExecutor`
and `TarballRecoveryExecutor`, as they will eventually be both
handled through the `MainRecoveryExecutor`.

Note that this is only a stub, and not a complete recovery executor.
The signature of methods, docstrings and their actual implementation
might/will change as part of future implementaions in this class.
This is only a kick-off for the final milestone.

References: BAR-577.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>