Add note to docs about compressing backups for network saving
If using backup compression and setting compression location to
server and format plain, backups will be compressed before leaving
the Postgres server and then decompressed on the client side
before ending in the final destination. This can be used if the
user wants to lower network usage, but keep the backups in
plain format, uncompressed.

References: BAR-609

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>