Fix error message when parsing invalid `--target-time` in `barman restore`
Previously, when parsing an invalid `--target-time` string in
`barman restore`,  the user encountered the following error:

    EXCEPTION: local variable 'parsed_target' referenced before assignment

This occurred because an exception in `load_datetime_tz` left
`parsed_target` undefined,  causing an unhandled exception during error
handling.

This commit ensures that the exception is properly caught and replaced
with a  more informative error, providing the user with a clear error
message.

References: BAR-627

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>