Fix issue in the docs of `custom_compression_magic`
The `custom_compression_magic` docs had a tip about how to configure
it when using the `xz` tool as a custom compression algorithm in
Barman.

However, the tip was leading the user to set that option like this:

    `custom_compression_magic = fd377a585a00`

While Barman expects the value of `custom_compression_magic` to be
prefixed with `0x`. This commit extends that tip to clarify the
expected value for the option in that specific example, i.e.
`0xfd377a585a00`.

References: BAR-625.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>