Clarify handling of Barman custom compression in the docs
Clarified some internal concepts regarding the usage of custom compression
in Barman. It states that Barman does not store matadata of compression
used for each WAL file. Instead it relies on the current server
configuration for compressing/decompressing. This means that the user
should always consider the current configuration and maintain it
consistent with the WALs in archive.

References: BAR-628

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>