Fix barman restore with --no-get-wal and --standby
Fixed an issue where Barman was removing the `pg_wal` directory after
recovery when `--no-get-wal` and `--standby-mode` were specified together.

The issue arises from Barman incorrectly filling the
recovery.conf/postgres.auto.conf file. When using standby mode with
no-get-wal, Barman moves WAL files directly to the `pg_wal` directory.
However, it was still setting `restore_command` to reference `pg_wal` as
an archiving source, which is redundant. Additionally, it was filling
`recovery_end_command` with instructions to remove `pg_wal`, making
recovery impossible.

This issue of Barman removing `pg_wal` was introduced in Barman 3.13.0.
However, using `--no-get-wal` and `--standby-mode` in prior versions
still have an inconsistent behavior, where Barman moves WALs to `pg_wal`
but references a non-existent `barman_wal` directory in both
`restore_command` and `recovery_end_command`. However in this case,
despite the inconsistencies, Postgres can still start, as the WALs are
accessible in `pg_wal`.

This is all due to the behavior of a standby during startup:

"At startup, the standby begins by restoring all WAL available in the
archive location, calling restore_command. Once it reaches the end of
WAL available there and restore_command fails, it tries to restore any
WAL available in the pg_wal directory. If that fails, and streaming
replication has been configured, the standby tries to connect to the
primary server and start streaming WAL from the last valid record found
in archive or pg_wal. If that fails or streaming replication is not
configured, or if the connection is later disconnected, the standby goes
back to step 1 and tries to restore the file from the archive again.
This loop of retries from the archive, pg_wal, and via streaming
replication goes on until the server is stopped or is promoted."

So, as long as we have one of the three (`restore_command`, WALs in
`pg_wal` or streaming replication), the startup should succeed.

That said, the fix introduced in this commit simply ignores filling any
recovery parameters if `--no-get-wal` and `--standby-mode` are in use,
as all required WALs will already be under the `pg_wal` directory.

Reported by: Boriss Mejias
References: BAR-630

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>