Implement relocation and renaming of backup.info files
Implements the relocation and renaming of the `backup.info` files.

Previously, these files were stored inside the backup directory alongside
the pgdata folder. With the introduction of support for backups in
immutable storage (also known as WORM—Write Once, Read Many), a new
location became necessary.

This commit relocates the backup.info file to a new <backup_directory>/meta
directory and renames it to <backup_id>-backup.info. This change enables
support for backups in WORM storages, as this directory must always
reside in a mutable partition.

References: BAR-650

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>