Make clearer that reusing the destination directories is dangerous
By default, Barman performs what is known as a "delta restore". That
means that when the destination directory (for base or tablespaces)
already exist in the recovery host, `barman restore` will simply apply
a delta between the files that exist in the recovery host, and the
files that exist in the Barman host.

While that is a nice feature, that behavior is dangerous. For example:
if you run `barman restore` to restore your backup targeting the same
directories where your Postgres instance is running, you can end up
losing data and/or corrupting your database cluster.

This commit makes it clearer in the docs that reusing the destination
directories can be dangerous.

Note: we are studying the possibility of turning the delta restore
into an optional behavior of `barman restore`. It might be safer to
error out in case the destination directories are not empty, but still
provide the delta restore as an optional behavior. For now, we won't
mention anything about "delta restore" in the docs to avoid confusion.

References: BAR-654.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>