Add VolatileBackupInfo class for in-memory backup updates
Introduced the `VolatileBackupInfo` class to support temporary,
in-memory updates to backup information without persisting changes
to disk. This class is designed to facilitate operations like
decryption, decompression and combination of backups with custom
tablespace mappings when creating and restoring a backup.

Add unit tests for this new class.

References: BAR-611

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>