Fix local incremental backup restore
When restoring an incremental backup locally, Barman fails with the
following error:

ERROR: Destination directory '/home/gustavo/Documents/to-recover/' must
be empty.

The error still appears even if the directory is cleaned up or completely
removed before the restore.

The problem is that Barman is shooting its own foot by moving directories
inside the destination directory before performing the is-empty check.

Unlike what happens when copying files with Rsync, the `_backup_copy`
method of the `IncrementalRecoveryExecutor` class does not ignore copying
certain directories like `pg_wal`.

This was not a problem until Barman 3.13.0, when the `--staging-wal`
option was introduced and along with that it started checking that the
WALs destination directory is empty.

So what caused the issue was:
1. Barman moves `pg_wal` (which has subdirectories, but not WAL files)
   to destination directory;
2. Before copying the actual WALs it checks if the destination exists and
   is empty. In case of incremental backup restore, it will not be, so it
   errors out.

That said, this commit fixes the issue by ignoring directories such as
`pg_wal` when moving files during incremental backup restore,
following the same reasoning as Rsync.

References: BAR-655

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>