Add `encryption` Field to BackupInfo class
This commit introduces a new 'encryption' field to the BackupInfo
class, which is critical for enhancing backup management. The new
field will store the encryption option [None, gpg] necessary for
decrypting backups during the `barman restore` process. By including
this information in the backup.info file, we ensure that encrypted
backups can be properly identified and decrypted when performing
restores.

Fixes the unit tests that were broken by this change, which introduced
the new field `encryption` in the BackupInfo class.

References: BAR-691

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>