Fix barman-cloud-backup not recycling temporary part files
Fixed a barman-cloud-backup problem where it was not deleting temporary
part files after being uploaded to the cloud, leading to disk space problems
over time.

- Summary

The problem happens because there is a broken logic in the
`async_upload_part` method of the `CloudInterface`, where it was
mistakenly instantiating a duplicate part file as a temporary file,
causing the oldest file to still persist in the temporary directory
while only the new one was recycled after the upload. The broken code is
only executed in Python >= 3.12, explaining why it remained unnoticed up
to now.

- Detailed explanation

Barman used to execute the following logic:

1. Create the part files with a specific name (something prefixed with `barman-upload-*`)
2. When it got to the code removed in this commit it could follow two paths:
    A. If the file is already a named-temporary file, it just keeps
        using the same file reference from that on.
    B. If not, it instantiates a new named-temporary file and copies the
        content of the current file to it.

The problem is that, in case B, when creating a new named-temporary file,
the random file gets a random name, different from that of the file already
created. This will result in two files with the same content, but different
names, living in the temporary directory.

What will happen from that on is that that only the random-named file
will be referenced, consequently resulting in it also being the only one
removed after the upload is finished.

Essentially, case B is wrong and should NEVER be executed, as there is no
reason to duplicate the existing file

Why does it only happen if Python version >= 3.12?

Case B should NEVER get executed as, after investigation, there will never be a
case where something other than a named-temporary file is passed to
`async_upload_part`. However, the condition which determined if the file was
a named-temporary file is the following:

`if hasattr(body, "name") and hasattr(body, "delete") and not body.delete:`

From Python >= 3.12, named-temporary file objects no longer have the `delete`
attribute, causing the code to mistakenly think it is other kind of object
and fall into case B.

- Solution

Just removed the whole condition of duplicating files as it was proved that
nothing other than a named-temporary file object will ever arrive to the
`async_upload_part` method.

References: BAR-674

Closes: #1076

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>