Add command to list all server subprocesses
This commit adds the new `list-processes` command that lists the active
subprocesses of a server. It displays the PID and task name
for each of the active subprocesses.

The implementation takes advantage of the already existing code
of the `ProcessManager` class: we only had to add more lock classes
to the supported task list, and as a consequence, provide a regex pattern
to match their lock file names.

Note: `barman list-processes` is a command at the server level. With that
in mind, this commit only adds support for listing processes which are also
at the server level. Processes like `barman cron` and `barman config-update`,
which are global, for example, were left out of scope on purpose.

References: BAR-664

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>