Use Ubuntu 22.04 runner for public CI, and exclude Python 3.6
This commit is a workaround so we have our public CI workflow back to
normal working.

On April 15th, 2025, GitHub dropped the Ubuntu 20.04 runner, and our
public CI workflow started failing because of that. If we want to keep
using the GitHub Ubuntu runners, we need to use either 22.04 or 24.04.

However, neither Ubuntu 22.04 nor Ubuntu 24.04 runners have Python 3.6
available, and we need to support Python 3.6 in Barman as that's the
default Python version for RHEL 8 and SLES 15, which are supported
platforms.

For now, as a workaround, we are excluding Python 3.6 from the matrix
and using the Ubuntu 22.04 runner. Later we might have a proper solution
for this issue.

References: BAR-698.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>