Fix regression when using `barman-cloud-backup` as a hook script
Barman 3.13.2 introduced a fix for handling WORM environments, which
involved relocating the `backup.info` file out of the base backup
directory. However, as a side effect, that introduced a regression
when using `barman-cloud-backup` as a hook script: at that time, that
code path was not properly updated to reflect the changes in the location
of `backup.info`.

That regression was causing users to face errors like this in that
situation:

 ERROR: Barman cloud backup exception: [Errno 2] No such file or directory: '/home/vagrant/barman_home/postgres/base/20250422T173749/backup.info'

This commit fixes that issue by making `barman-cloud-backup` aware of the
`backup.info` path, when running as a hook script in the Barman server.

As the `barman-cloud-backup` knows nothing about the Barman server, and
it is not supposed to instantiate objects that belong to the server, we
made the script aware of that file by introducing a new environment
variable in the script -- `BARMAN_BACKUP_INFO_PATH` -- which is set by
`barman backup` with the path to the `backup.info` of the backup that
triggers the hook. That way we can reliably get the path, and do not
depend on instances of objects that belog to the Barman server.

References: BAR-696

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>