Implement GPG encryption of tar base backups and tablespaces
This commit introduces support for GPG encryption of base backups and
tablespaces when using the tar format. It is the initial
implementation of the backup encryption feature, limited for now to only
Postgres backup method with tar backups. This limited approach makes
it easier to deliver and build upon the feature incrementally.

This implementation includes classes responsible for performing encryption
(and later decryption), as well as an `EncryptionManager` class that
handles validation and instantiation of encryption handlers (currently
only GPG). All validation logic is centralized in the manager, which will
facilitate its reuse when implementing the ecryption check in `barman check`.

For now, encryption failures are detected at the end of the backup process,
also marking the backup as failed. Once the encryption check is implemented,
this check will happen before the backup starts, avoiding later fails.

References: BAR-683

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>