Add new check for encryption in barman check
Added a new check for the configured encryption in the `barman check`
command. It basically validates the configuration and attempts a simple
encryption with the respective encryption class. This will make Barman
fail early in case something is misconfigured, avoiding the user waiting
until the end of the backup to find out encryption does not work.

References: BAR-702

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>