Add `terminate-process` command to terminate an active subprocess from a server
This commit introduces a new `terminate-process` command that allows users
to terminate an active subprocess for a given server by specifying its
task name. The command calls `server.kill` to terminate the process, which
implicitly allows the termination of a task only if it's active and owned
by the specified server.

References: BAR-665

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>