Add documentation for the new `list-processes` command
This updates the documentation to describe the usage of the
new `list-processes` command and to include the reference to it.

Note: this commits adds a new page to the manual. Thus, we
also had to add that entry to the `setup.py`, so it is shipped in
the Python package. Also, for that to work, we had to pre-compile
the new man page, so we avoid failures when building the Python
package.

References: BAR-664

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>