Add documentation on the new `barman terminate-process` command
This adds a new documentation page to the manual and updates the
references regarding the new `terminate-process` command, used to
terminate an active specified subprocess for a given barman server.
This commit also updates `setup.py` accordingly and adds the built
new man page to avoid failures when packaging.

References: BAR-665

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>