Remove pin from version of boto3
In a previous release of Barman we pinned the version of boto3
to a version prior to 1.36. This was because a breaking change
was introduced in that version.

After further investigation, and seeing that object store
facilities like MinIO have addressed the issues with boto3
>= 1.36.0, we will remove the pin and use the latest version.

This brings more compatibility with FIPS in the Barman cloud
scripts.

References: BAR-637

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>