Clarify path_prefix usage to reflect actual behavior
The `path_prefix` option is currently used to locate all types of
binaries in Barman, including rsync, compression tools, PostgreSQL,
and more. However, the documentation inaccurately states that it
only needs to point to the bin directory for the appropriate
`PG_MAJOR_VERSION`.

This commit changes the description of this configuration option
to clarify its use.

References: BAR-715

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>