Add a section about local recovery
We didn't document very thoroughly the steps that were needed
in case of doing a local recovery. This commit is a first attempt
at filling the gap in the documentation.

We clearly state the two options, running postgres with the barman
user, or with the postgres user, and things the person restoring
has to keep in mind.

References: BAR-676

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>