Add documentation regarding encrypted backups in WORM environments
When encrypting backups in WORM environments, the user has to take
into account that the original backup files will need to be
encrypted after the backup finishes.

The requirement is that the grace period has to be long enough
so that after the original tarball's copy finishes, there is time
to finish copying the other tarballs (if there are multiple
tablespaces involved) and encrypting the former tarball.

This is because encryption creates a new file, the tarball that is
encrypted, and once finishing the encryption, the original tarball
is deleted.

References: BAR-694

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>