Fix usage of barman-wal-restore with --keep-compression with gzip,bzip2 and pigz
Fixed an issue in `barman-wal-restore` where, when trying to restore WALs
compressed with `gzip`, `bzip2` or `pigz` while having `--keep-compression`
specified, lead to unexpected errors.

The root of the error is that all these algorithms are invoked via a
subprocess call, but the name of the process to be spawned was being
taken from `config.compression`, which does not make sense as
`config.compression` will always be `None` when `--keep-compression` is
set. This was most likely an oversight of some refactoring that took place
in compression modules when implementing new features.

References: BAR-722

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>