Pin pip version to 25.0.1 due to a pip-tools bug
Due to the following pip issue: jazzband/pip-tools#2176 we are going
to pin pip version to 25.0.1.

This is a temporary workaround. Once the issue is fixed in pip-tool,
we can revert this patch.

Note: caching was also removed because it part of the issue. We were
getting several permission denied errors in the caching mechanism.
After removing caching and pinning the pip version, the CI workflow
started working fine again.

References: bar-726

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>
Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>