Validate options passed for servers and models through `config-update`
This adds a new method `_validate_section` to `ConfigChangesProcessor`.
The method contains the validations that need to be performed before
a config update, including the validations previously performed
on `receive_config_changes` and new validations performed for each
key/value through its parser. The method returns a bool according
to the validation results and additional information so the config
changes can proceed.

This commit also alters the behavior of `receive_config_changes` and
of the `config-update` command. Now, any invalid changes passed to the
command will cause it to error out, as opposed to ignoring the invalid
sections and processing the valid ones.

References: BAR-142

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>