Automatically switch a WAL when the archiving check fails
When the archiving check fails, users are advised to
manually switch a WAL file and run `barman cron` (as per docs).
This is because Barman relies primarily on the contents of the `xlog.db`.
If no WALs are present there, it assumes the archiving is not working
correctly and advises users to switch a WAL and run cron to see if
content arrives.

This is a process that can be automated, as Barman has the ability to
identify an empty `xlog.db` and automatically switch and archive WAL
files internally.

That said, this commit implements the following logic:

If Barman notices there's nothing in the `xlog.db` and no previous
backups, instead of giving an error it will:

1. Check if there are WALs in the incoming/streaming directories. If
   there are, run internal `barman archive-wal` command to archive
   them.

2. If there are no WALs in the incoming/streaming directories, it will
   run the internal `barman switch-wal` command to switch to a new WAL
   file and try to archive it as soon as it arrives. The timeout for
   the WAL file to arrive is <check_timeout> / 3.

3. If Barman still wasn't able to archive anything, then we consider the
   check as failed.

The idea is to provide a better user experience by not making users
have to run such commands manually whenever configuring a new server or
when Barman has `xlog.db` empty for some reason.

References: BAR-718

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>