Refactor MainRecoveryExecutor to use operation classes
The `MainRecoveryExecutor` as it was (having handler methods for
each operation) would be having too much responsibility and handling
lots of scenarios alone. This would difficult testing and general
understanding of the code.

To better organize the code, the `MainRecoveryExecutor` will now be a
general class which coordinates the operations (created in the previous
commit) to be run on the backup (copy, decompression, combination and
decryption).

This refactor will allow us to have a cleaner code, with each operation
being handled by its own class, making it easier to test and maintain.

References: BAR-758

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>