Allow incremental backups with compression
Barman used to block the user from attempting to take compressed
incremental backups. If an incremental backups was attempted with either
compression enabled or upon a compressed parent backup, Barman would
error out.

While `pg_basebackup` allows that, we blocked that situation when the
incremental backups were initially implemented because we had no way of
performing a restore over compressed and incremental backups.

This commit removes that restriction, as with the new structure provided
by the new `MainRecoveryExecutor` such restores are now possible.

References: BAR-583

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>