Add _post_recovery_cleanup method to RecoveryOperation
Added a new `_post_recovery_cleanup` method to the `RecoveryOperation` class.
This is a helper method used to remove unwanted files from the destination
directory after the operation is completed, such as contents inside the
`pg_wal`, `pg_log`, the `postmaster.pid` file, etc.

This method is useful for certain operations that direct its output
directly to the final destination, without the means to ignore specific
files.

References: BAR-792

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>