Create logic for handling tablespaces symbolic links
`pg_combinebackup` requires the symbolic links of the backups
tablespaces to be correct previous to combining them, otherwise
tablespaces are ignored.

That said, we need to ensure that the symbolic links
are created, mainly after operations such as `RsyncCopyOperation`
and `DecompressOperation`, which are the only ones that can preceed
a `CombineOperation` in the recovery process.

This new method is intened to replace the old `_prepare_tablespaces` method,
which performed a similar function but created the links before the copy
process and was only called once per recovery process, essentially dealing
only with the final destination of the tablespaces, not with the case where
they are in the staging area.

That said, this commit adds a new method `_link_tablespaces` to the
`RecoveryOperation` class, which will handle that and can be reused
by operations when needed.

The method creates a symbolic link from `PGDATA/pg_tblspc/<oid>` to
`<tablespace_path>`. `<tablespace_path>` will either point to:

1. The final tablespace location, if this is the last operation;

2. The tablespace location in the staging area, if more operations will follow.

Intended usage:

- After the `RsyncCopyOperation` or `DecompressOperation` is done,
  call `_link_tablespaces` to create the symbolic links for the
  tablespaces, as a `CombineOperation` might come next. If it's the
  last operation, it will do the same as the old `_prepare_tablespaces`.
- `DecryptOperation` does not need it because it will always be followed
  by a `DecompressOperation` and will never be the last operation.
- `CombineOperation` does not need to call it either, as it
  `pg_combinebackup` already handles the symbolic links when relocating.

References: BAR-800

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>