Extend CombineOperation to run remotely
Extended the `CombineOperation` class to support remote execution of the
`pg_combinebackup` command. To achieve this, when `staging_location` is
`remote`, we build a new `Command` class which executes `pg_combinebackup`
with the correct arguments prefixed by the remote host's SSH command.

 `_fetch_remote_status` also had to be changed to support remote execution.
 It is responsible for checking if `pg_combinebackup` is installed and
returning its full path and version.

References: BAR-767

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>