Fail gracefully when an unexpected field is found in the backup metadata
Adds a proper message when an `AttributeError` is raised in case
Barman loads the `backup.info` file and an unexpected field is
present in it.

This happens, for example, when a user tries to restore a backup
with an older version than the one it was used to create the backup
in the first place. The backup can have a field that an older
Barman version cannot recognize. We must tell the user that this
can't happen and Barman must be upgraded.

References: BAR-763

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>